Add files via upload
This is an extension of already made speech to text model but there is a working GUI made. This facilitate the conversion and also allows to add new feature such as reading the text aloud in male or female voice, allows to increase or decrease volume and the speed at which the text is read. There is a feature to save the text in a .txt file and to clear the screen.

Made by Abhyuday and Vansh